fix `error c2036: 'void *': Unknown size` on MSVC

+++ rabbyt/anim_sys.h
@@ -62,7 +62,7 @@ enum {
             (out)[0] = (slot)->local;\
             break;\
         default:\
-            (out)[0] = ((float*)((slot)->base[0] + (slot)->offset))[0];\
+            (out)[0] = ((float*)((int)(slot)->base[0] + (int)(slot)->offset))[0];\
             break;\
     }}while(0)
